+++ src/feature/cards/cards.service.ts
@@ -1,20 +1,28 @@
 import { Injectable } from '@nestjs/common';
 import _ from 'lodash';
 import moment from 'moment';
+import { ApiService, SettingsDto } from '../../shared/api';
 import { CardMapper, CardService, CardTemplate } from '../../shared/game';
 import { CardDocument } from './ui/cards.document';
 
 @Injectable()
 export class CardsService {
-  constructor(private cardService: CardService) {}
+  constructor(
+    private apiService: ApiService,
+    private cardService: CardService,
+  ) {}
 
   async getCardDocuments(): Promise<CardDocument[]> {
     const cardTemplates = await this.cardService.getAllCardTemplates();
+    const settings = await this.apiService.fetchSettings();
 
-    return this.mapDocuments(cardTemplates);
+    return this.mapDocuments(cardTemplates, settings);
   }
 
-  mapDocuments(cardTemplates: CardTemplate[]): CardDocument[] {
+  mapDocuments(
+    cardTemplates: CardTemplate[],
+    settings: SettingsDto,
+  ): CardDocument[] {
     const now = moment().unix();
 
     return _.chain(cardTemplates)
@@ -22,13 +30,24 @@ export class CardsService {
         const cards = [];
 
         for (const distribution of cardTemplate.distributions) {
-          for (const level of [1, 2, 3, 4]) {
+          for (const level of _.range(1, 10)) {
+            const xp = CardMapper.mapToCardsToLevel(
+              cardTemplate.id,
+              level,
+              distribution.editionNumber,
+              cardTemplate.rarity,
+              undefined,
+              distribution.gold,
+              settings,
+            );
+
             cards.push(
               CardMapper.mapToCard(
                 cardTemplate,
                 level,
                 distribution.editionNumber,
                 distribution.gold,
+                xp,
               ),
             );
           }
@@ -50,14 +69,17 @@ export class CardsService {
           level: card.level,
           mana: card.stats.mana,
           name: card.name,
+          power: card.power,
           rarity: card.rarity,
           rarityNumber: card.rarityNumber,
           role: card.type,
           splinter: card.splinter,
+          xp: card.xp,
         };
 
         return document;
       })
+      .filter((it) => !!it.xp)
       .value();
   }
 }