+++ src/feature/cards/ui/cards.document.ts
@@ -16,8 +16,10 @@ export class CardDocument extends DocumentDto {
   readonly level: number;
   readonly mana: number;
   readonly name: string;
+  readonly power: number;
   readonly rarity: string;
   readonly rarityNumber: number;
   readonly role: string;
   readonly splinter: string;
+  readonly xp: number;
 }