+++ src/util/ui/index.ts
@@ -1,1 +1,2 @@
 export * from './imageLabelCell';
+export * from './shared-columns';